refactor: move non-basic utility code from utils.lua to dedicated files
